Домашна 4/backend/news-scraper-service/requirements.txt
apscheduler
requests
beautifulsoup4~=4.12.3
aiohttp~=3.10.11
uvicorn
fastapi~=0.115.6
sqlalchemy~=2.0.34
psycopg2-binary
numpy==1.26.4
pymupdf
googletrans==4.0.0-rc1
python-docx